Extract custom hook
- Move new hook `useSideEffect` and files related to performing
  actions to separate directory inside `pages` directory. This is
  because all these are used both in the swap page and in the swap
  list.
