+++ src/pages/SwapList/SwapRow/SwapRow.tsx
@@ -1,25 +1,27 @@
 import { CircularProgress, TableCell, TableRow } from "@material-ui/core";
 import { makeStyles } from "@material-ui/styles";
-import React, { useEffect, useReducer } from "react";
+import React, { useReducer } from "react";
 import { RouteComponentProps, withRouter } from "react-router-dom";
 import { EmbeddedRepresentationSubEntity } from "../../../../gen/siren";
-import executeAction from "../../../api/executeAction";
 import { Asset, Properties, toMainUnit } from "../../../api/swapTypes";
 import ActionButton from "../../../components/ActionButton";
 import Dialog from "../../../components/Dialog";
 import ExternalLink from "../../../components/ExternalLink";
-import LedgerActionDialogBody from "../LedgerActionDialogBody";
-import SirenActionParametersDialogBody from "../SirenActionParametersDialogBody";
-import SwapStatusIcon from "../SwapStatusIcon";
 import {
   actionButtonClicked,
-  actionFailed,
-  actionSuccessful,
   closeLedgerActionDialog,
   closeSirenParametersDialog,
   sirenParameterDialogSubmitted
-} from "./events";
-import { ActionExecutionStatus, initialState, reducer } from "./reducer";
+} from "../../actions/events";
+import {
+  ActionExecutionStatus,
+  initialState,
+  reducer
+} from "../../actions/reducer";
+import useSideEffect from "../../actions/useSideEffect";
+import LedgerActionDialogBody from "../LedgerActionDialogBody";
+import SirenActionParametersDialogBody from "../SirenActionParametersDialogBody";
+import SwapStatusIcon from "../SwapStatusIcon";
 import SwapId from "./SwapId";
 
 interface AssetCellProps {
@@ -54,25 +56,7 @@ function SwapRow({ swap, history, reload }: SwapRowProps) {
     dispatch
   ] = useReducer(reducer, initialState);
 
-  useEffect(() => {
-    if (!sideEffect) {
-      return;
-    }
-
-    switch (sideEffect.type) {
-      case "reloadData": {
-        reload();
-        return;
-      }
-      case "executeAction": {
-        executeAction(sideEffect.payload.action, sideEffect.payload.data).then(
-          response => dispatch(actionSuccessful(response)),
-          error => dispatch(actionFailed(error))
-        );
-        return;
-      }
-    }
-  }, [sideEffect, reload]);
+  useSideEffect(reload, dispatch, sideEffect);
 
   const classes = useStyles();
