+++ src/pages/actions/reducer.spec.ts
@@ -1,5 +1,5 @@
 import { AxiosResponse } from "axios";
-import { Action } from "../../../../gen/siren";
+import { Action } from "../../../gen/siren";
 import {
   actionButtonClicked,
   actionSuccessful,
